+++ pysyringe/__init__.py
@@ -1,1 +1,5 @@
 __version__ = "1.5.3.dev3+g4027b3ceb.d20260210"
+
+from pysyringe.container import Container, Provide
+
+__all__ = ["Container", "Provide"]